move identity and queries into iparty
