+++ src/party/cloud/cloud-party.js
@@ -1,8 +1,7 @@
-const dataparty_crypto = require('@dataparty/crypto')
 const debug = require('debug')('dataparty.cloud.party')
 
-const Qb = require('./qb.js')
-const Query = require('./query.js')
+const Qb = require('../qb')
+const Query = require('../query')
 const IParty = require('../iparty')
 const RestComms = require('../../comms/rest-comms')
 
@@ -23,127 +22,8 @@ class CloudParty extends IParty {
       cache: this.cache
     })
 
-    this._actor = {id: undefined, type: undefined}
-    this._actors = []
-    this._identity = undefined
-    
-    //if(uri[uri.length-1] != '/'){ uri = uri+'/' }
-
     this.rest = new RestComms({config: this.config, party: this})
-  }
-
-  get comms(){
-    return this.rest
-  }
-
-  /**
-   * @method
-   */
-  hasIdentity(){
-    return this._identity != undefined
-  }
-
-  /**
-   * @method
-   */
-  hasActor(){
-    return this.actor != undefined
-  }
-
-  /**
-   * @type {module:dataparty/Types.Identity}
-   */
-  get identity(){
-    if (!this.hasIdentity()){ return undefined }
-    return this._identity.toJSON(false)
-  }
-
-  /** @type {IdObj} */
-  get actor(){
-    if (this.actors && this.actors[0]){
-
-      return this.actors[0]
-
-    } else if (this._actor.id && this._actor.type){
-
-      return this._actor
-
-    }
-
-    return undefined
-  }
-
-  /** @type {IdObj[]} */
-  get actors(){
-    return this._actors
-  }
-
-  set actors(value){
-    this._actors = value
-
-    const primaryActor = this.actor
-
-    if (!primaryActor){
-     return
-    }
-
-    this._actor.id = primaryActor.id
-    this._actor.type = primaryActor.type
-
-    const path = 'actor'
-    this.config.write(path, this._actor)
-  }
-
-  /**
-   * @method
-   */
-  async encrypt(data, to){
-    const msg = new dataparty_crypto.Message({msg: data})
-    await msg.encrypt(this._identity, to.key)
-
-    return msg
-  }
-
-  /**
-   * @method
-   */
-  async decrypt(reply, expectedSender, expectClearTextReply = false){
-    // if reply has ciphertext & sig attempt to decrypt
-    if (reply.enc && reply.sig) {
-      const msg = new dataparty_crypto.Message(reply)
-
-      const replyContent = await msg.decrypt(this._identity)
-
-      const publicKeys = dataparty_crypto.Routines.extractPublicKeys(msg.enc)
-
-      debug(`publicKeys.sign - ${publicKeys.sign}`)
-
-      if (publicKeys.sign !== expectedSender.key.public.sign ||
-          publicKeys.box !== expectedSender.key.public.box) {
-        throw new Error('TrustFail: reply is not from service')
-      }
-
-      debug('decrypted reply ->', JSON.stringify(replyContent))
-
-      if (replyContent.error) {
-        debug('call failed ->', replyContent.error)
-        throw replyContent.error
-      }
-
-      return replyContent
-
-    } else if (expectClearTextReply && !reply.error) {
-
-      return reply
-
-    }
-
-    if (reply.error) {
-      debug('call failed ->', reply.error)
-      throw reply.error
-    }
-
-    throw new Error('TrustFail: reply is not encrypted')
+    this.comms = this.rest
   }
 
   /**
@@ -166,54 +46,12 @@ class CloudParty extends IParty {
   async start(){
     debug('start')
     await super.start()
-    await Promise.all([
-      this.loadIdentity(),
-      this.loadActor(),
-    ])
 
     await this.rest.start()
   }
 
   async stop(){
     await this.rest.stop()
-  }
-
-  /**
-   * @method
-   */
-  async loadIdentity(){
-    const path = 'identity'
-    const cfgIdenStr = this.config.read(path)
-
-    if (!cfgIdenStr){
-      debug('generated new identity')
-      this._identity = new dataparty_crypto.Identity({id: 'primary'})
-      await this.config.write(path, this._identity.toJSON(true))
-    } else {
-      debug('loaded identity')
-      this._identity = dataparty_crypto.Identity.fromString(JSON.stringify(cfgIdenStr))
-    }
-  }
-
-  async resetIdentity(){
-    const path = 'identity'
-    await this.config.write(path, null)
-    await this.loadIdentity()
-  }
-
-  /**
-   * @method
-   */
-  async loadActor(){
-    const path = 'actor'
-    const localActorObj = this.config.read(path)
-
-    if (!localActorObj){ return }
-
-    this._actor.id = localActorObj.id
-    this._actor.type = localActorObj.type
-
-    debug('loaded actor', this._actor)
   }
 
 
@@ -252,7 +90,7 @@ class CloudParty extends IParty {
   /**
    * @method
    */
-  update (...msgs) {
+  async update (...msgs) {
     return this.qb.modify(msgs, 'update')
   }
 
@@ -277,7 +115,7 @@ class CloudParty extends IParty {
   /**
    * @method
    */
-  create (type, ...msgs) {
+  async create (type, ...msgs) {
     return this.qb.create(type, msgs)
   }
 
@@ -299,7 +137,7 @@ class CloudParty extends IParty {
   /**
    * @method
    */
-  remove (...msgs) {
+  async remove (...msgs) {
     return this.qb.modify(msgs, 'remove')
   }
 }
