+++ src/party/iparty.js
@@ -1,5 +1,7 @@
 const debug = require('debug')('dataparty.iparty')
+const dataparty_crypto = require('@dataparty/crypto')
 
+const Query = require('./query.js')
 const IDocument = require('./idocument')
 const DocumentFactory = require('./document-factory')
 const LokiCache = require('./loki-cache.js') // insert | populate cache
@@ -10,10 +12,15 @@ const LokiCache = require('./loki-cache.js') // insert | populate cache
  * @interface
  */
 class IParty {
-  constructor({config, cache, model, factories, documentClass}){
+  constructor({config, cache, comms, model, factories, documentClass}){
     this.config = config
     this.cache = cache || new LokiCache()
+    this.comms = comms
     
+    this._actor = {id: undefined, type: undefined}
+    this._actors = []
+    this._identity = undefined
+
     /**
      * @member factory 
      * @type {DocumentFactory} */
@@ -31,6 +38,11 @@ class IParty {
       await this.cache.start()
     }
 
+    await Promise.all([
+      this.loadIdentity(),
+      this.loadActor(),
+    ])
+
     debug('\tDocument Validators', this.factory.getValidators())
     debug('\tDocument Classes', this.factory.getTypes())
   }
@@ -38,32 +50,44 @@ class IParty {
   /**
    * @method
    */
-  createDocument(type, data, id){
+  async createDocument(type, data, id){
     let Type = this.factory.getFactory(type)
 
     return Type.create(this, {data, type, id})
   }
 
-  create (type, ...msgs) {
-    throw new Error('Not Implemented')
+  /**
+   * @method
+   */
+  async create (type, ...msgs) {
+    return this.qb.create(type, msgs)
   }
 
-  remove (...msgs) {
-    throw new Error('Not Implemented')
+
+  /**
+   * @method
+   */
+  async remove (...msgs) {
+    return this.qb.modify(msgs, 'remove')
   }
 
-
-  update (...msgs) {
-    throw new Error('Not Implemented')
+  // takes modified json msgs & writes to backend, resolves to new stamps
+  // requires type & id
+  /**
+   * @method
+   */
+  async update (...msgs) {
+    return this.qb.modify(msgs, 'update')
   }
-  
+
+  /**
+   * @method
+   */
   find () {
-    throw new Error('Not Implemented')
+    return new Query(this.qb, this.factory)
   }
 
-  get comms(){
-    return undefined
-  }
+ 
 
   /**
    * @method
@@ -97,6 +121,156 @@ class IParty {
    */
   get types(){
     return this.factory.getFactories()
+  }
+
+  /**
+   * @type {module:dataparty/Types.Identity}
+   */
+  get identity(){
+    if (!this.hasIdentity()){ return undefined }
+    return this._identity.toJSON(false)
+  }
+
+  /** @type {IdObj} */
+  get actor(){
+    if (this.actors && this.actors[0]){
+
+      return this.actors[0]
+
+    } else if (this._actor.id && this._actor.type){
+
+      return this._actor
+
+    }
+
+    return undefined
+  }
+
+  /** @type {IdObj[]} */
+  get actors(){
+    return this._actors
+  }
+
+  set actors(value){
+    this._actors = value
+
+    const primaryActor = this.actor
+
+    if (!primaryActor){
+     return
+    }
+
+    this._actor.id = primaryActor.id
+    this._actor.type = primaryActor.type
+
+    const path = 'actor'
+    this.config.write(path, this._actor)
+  }
+
+  /**
+   * @method
+   */
+  hasIdentity(){
+    return this._identity != undefined
+  }
+
+  /**
+   * @method
+   */
+  hasActor(){
+    return this.actor != undefined
+  }
+
+
+
+    /**
+   * @method
+   */
+  async loadIdentity(){
+    const path = 'identity'
+    const cfgIdenStr = this.config.read(path)
+
+    if (!cfgIdenStr){
+      debug('generated new identity')
+      this._identity = new dataparty_crypto.Identity({id: 'primary'})
+      await this.config.write(path, this._identity.toJSON(true))
+    } else {
+      debug('loaded identity')
+      this._identity = dataparty_crypto.Identity.fromString(JSON.stringify(cfgIdenStr))
+    }
+  }
+
+  async resetIdentity(){
+    const path = 'identity'
+    await this.config.write(path, null)
+    await this.loadIdentity()
+  }
+
+  /**
+   * @method
+   */
+  async loadActor(){
+    const path = 'actor'
+    const localActorObj = this.config.read(path)
+
+    if (!localActorObj){ return }
+
+    this._actor.id = localActorObj.id
+    this._actor.type = localActorObj.type
+
+    debug('loaded actor', this._actor)
+  }
+
+    /**
+   * @method
+   */
+  async encrypt(data, to){
+    const msg = new dataparty_crypto.Message({msg: data})
+    await msg.encrypt(this._identity, to.key)
+
+    return msg
+  }
+
+  /**
+   * @method
+   */
+  async decrypt(reply, expectedSender, expectClearTextReply = false){
+    // if reply has ciphertext & sig attempt to decrypt
+    if (reply.enc && reply.sig) {
+      const msg = new dataparty_crypto.Message(reply)
+
+      const replyContent = await msg.decrypt(this._identity)
+
+      const publicKeys = dataparty_crypto.Routines.extractPublicKeys(msg.enc)
+
+      debug(`publicKeys.sign - ${publicKeys.sign}`)
+
+      if (publicKeys.sign !== expectedSender.key.public.sign ||
+          publicKeys.box !== expectedSender.key.public.box) {
+        throw new Error('TrustFail: reply is not from service')
+      }
+
+      debug('decrypted reply ->', JSON.stringify(replyContent))
+
+      if (replyContent.error) {
+        debug('call failed ->', replyContent.error)
+        throw replyContent.error
+      }
+
+      return replyContent
+
+    } else if (expectClearTextReply && !reply.error) {
+
+      return reply
+
+    }
+
+    if (reply.error) {
+      debug('call failed ->', reply.error)
+      throw reply.error
+    }
+
+    throw new Error('TrustFail: reply is not encrypted')
   }
 }
